% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrapDist}
\alias{bootstrapDist}
\title{bootstrapDist}
\usage{
bootstrapDist(
  co_gr,
  B = 100,
  mapping_fun = "k",
  chr = NULL,
  group_prefix,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{co_gr}{GRanges the crossover counts for each marker interval across all
samples. Returned by \code{countCO}}

\item{B}{integer the number of sampling times}

\item{mapping_fun}{character default to "k" (kosambi mapping function). It 
can be one of the mapping functions: "k","h"}

\item{chr}{character the specific chromosome to test or NULL indicates all
chrs in the GRanges object will be summed together.}

\item{group_prefix}{the prefix for each group that we need to generate 
distributions for}

\item{BPPARAM, }{the bpparam backend registered for parallel computing}
}
\value{
lists of numeric genetic distances for multiple samples
}
\description{
Generating distribution of sample genetic distances
}
\details{
It takes the crossover counts for samples in multiple groups that is returned
by `countCO`. It then draws samples from a group with replacement and 
calculate the distribution of relevant statistics.
}
\author{
Ruqian Lyu
}
