% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-coc.R
\name{calCoC}
\alias{calCoC}
\title{calculate the  coefficient of coincidence for grouped samples}
\usage{
calCoC(count_co_gr, n_intervals = 11, ref_genome = "mm10", group_prefix = NULL)
}
\arguments{
\item{count_co_gr, }{the crossover count genomic ranges object}

\item{n_intervals, }{int, the number of intervals to subset each chromosome}

\item{ref_genome, }{character, the reference genome}

\item{group_prefix, }{character vector, the prefix for grouping samples in 
sample names.}
}
\value{
a data.frame with CoC calculated for interval pairs in each
chromosome
}
\description{
It takes the crossover count genomic range object and distribute the crossovers
to `n_intervals` per chromosome. The intervals along each chromosome are then
paired and the observed double crossover rate is calculated. The coc is then
derived by taking ratio of observed_db_co_rate over expected_db_co_rate.
}
\author{
Ruqian Lyu
}
