% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-genetic-dist.R
\name{calGeneticDist}
\alias{calGeneticDist}
\title{calGeneticDist}
\usage{
calGeneticDist(
  co_geno_gr,
  bin_size = NULL,
  mapping_fun = "k",
  ref_genome = "mm10",
  by_group = NULL,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{co_geno_gr}{GRange object, returned by \code{countCO}}

\item{bin_size}{The binning size for grouping marker intervals into bins. If not supplied,the
orginial marker intervals are returned with converted genetic distancens
based on recombination rate}

\item{mapping_fun}{The mapping function to use, can be one of "k" or "h" (kosambi or haldane)}

\item{ref_genome}{The reference genome name. It is used to fetch the chromosome Information
from UCSC database.}

\item{by_group, }{the unique prefix of sample names that are used for defining
different sample groups. If NULL all samples are assumed to be from one group}

\item{BPPARAM, }{the back-end of type bpparamClass}
}
\value{
GRanges object
GRanges for marker intervals or binned intervals with Haldane or Kosambi 
centiMorgans
}
\description{
Calculate genetic distances of marker intervals or bins
}
\details{
Given whether crossover happens in each marker interval, calculate the
recombination fraction in samples and then derive the Haldane or Kosambi
genetic distances via mapping functions
}
