% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{calGeneticMap}
\alias{calGeneticMap}
\title{Calculate genetic map length}
\usage{
calGeneticMap(gt_matrix_co_by_marker, alpha = 0.05)
}
\arguments{
\item{gt_matrix_co_by_marker}{data.frame, from \code{melt} the matrix returned by \code{detectCO}}
}
\value{
data.frame
data.frame for all markers with Haldane and Kosambi morgans calculated
}
\description{
Given whether crossover happens in every two marker interval, calculate the
recombination fraction and then derive the Haldane or Kosambi morgans.
}
