% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{change_missing}
\alias{change_missing}
\title{change 'missing' labels to \code{NA}}
\usage{
change_missing(s_gt, missing = "missing")
}
\arguments{
\item{s_gt, }{a column of labelled genotypes}

\item{missing, }{the string used for encoding missing values default to \code{missing}}
}
\description{
change 'missing' labels to \code{NA}
}
