% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coCount}
\alias{coCount}
\title{RangedSummarizedExperiment object containing the crossover counts across
samples for the list of SNP marker intervals}
\format{
An object of class \code{RangedSummarizedExperiment} with 3 rows and 10 columns.
}
\usage{
coCount
}
\description{
RangedSummarizedExperiment object containing the crossover counts across
samples for the list of SNP marker intervals
}
\keyword{datasets}
