% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comapr.R
\docType{package}
\name{comapr}
\alias{comapr}
\title{\code{comapr} package}
\description{
crossover inference package
}
\details{
See the README on
 \href{https://gitlab.svi.edu.au/biocellgen-private/comapr}{GitLab}
}
