% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{correct_ref}
\alias{correct_ref}
\title{Function for correct Markers}
\usage{
correct_ref(s_gt, change_to = "Het")
}
\arguments{
\item{s_gt}{a vector of genotype labels as returned by \code{label_gt}}

\item{change_to}{by default, the Home_ref is changed to \code{Het}, but it can be changed to any labels
such as \code{Fail} fi we want to exclude these wrong labels.}
}
\value{
a vector of genotype labels with Home_ref corrected to \code{Het} or \code{Fail} as
specified by the argument change_to.
}
\description{
If we see home ref markers in the samples, then something is wrong.
We can change all home_ref to hets
}
\keyword{internal}
