% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{count_cos}
\alias{count_cos}
\title{Count crossovers
count the number of crossovers for each sample, per chromosome, cumulatively}
\usage{
count_cos(s_gt, interval_df, chrs, chrPos, type = "bool")
}
\arguments{
\item{s_gt}{genotypes of a sample along markers on all chromosomes}

\item{type}{"counts", or "bool" to specify whether this function should return
vector of cumulative counts or vector of TRUE/FALSE}

\item{chr}{a vector of characters indicating the chromosome locations of these markers}
}
\description{
Count crossovers
count the number of crossovers for each sample, per chromosome, cumulatively
}
\author{
Ruqian Lyu
}
