% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{countGT}
\alias{countGT}
\title{countGT
count how many samples have genotypes calls across all markers
and count how many markers that each individual have genotypes}
\usage{
countGT(gt_matrix, plot = TRUE, interactive = FALSE)
}
\arguments{
\item{gt_matrix, }{the genotype matrix of marker by samples after correction by
funcion \code{correctGT}}

\item{plot, }{it determines whether a plot will be generated.}
}
\value{
A list of two elements including \code{n_markers} and \code{n_samples}
}
\description{
countGT
count how many samples have genotypes calls across all markers
and count how many markers that each individual have genotypes
}
\author{
Ruqian Lyu
}
