% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{detectCO}
\alias{detectCO}
\title{detect crossovers}
\usage{
detectCO(gt_matrix, prefix = "Sample_", chrs, chrPos, type = "bool")
}
\arguments{
\item{prefix}{what prefix to add to sample IDs}

\item{chrs}{the chromosomes for markers on the rows}

\item{type}{whether return boolean value indicating whether crossover happens or integer value
indicating how many crossovers happen from all preceding intevals}

\item{gt_matix}{correct genotype matrix returned by \code{correct_gt}}
}
\value{
a matrix of marker by samples with values indicating crossovers
}
\description{
detect crossovers between every two markers by examining change in genotype patterns
}
