% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat-geno-mt.R
\name{.change_missing}
\alias{.change_missing}
\title{change SNPs with genotype 'Fail' to \code{NA}}
\usage{
.change_missing(s_gt, missing = "Fail")
}
\arguments{
\item{s_gt, }{a column of labelled genotypes}

\item{missing, }{the string used for encoding missing values default to
\code{Fail}}
}
\value{
a vector of genotypes with Fail substituted by `NA`
}
\description{
change SNPs with genotype 'Fail' to \code{NA}
}
\details{
calculation.
}
\author{
Ruqian Lyu
}
\keyword{internal}
