% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHapState.R
\name{filterCOs}
\alias{filterCOs}
\title{Filter out doublet cells and uninformative SNPs}
\usage{
filterCOs(se, minSNP = 30, minlogllRatio = 200, bpDist = 100, maxRawCO = 10)
}
\arguments{
\item{se, }{the SummarizedExperiment object that contains the called haplotype
state matrix in the assay field and haplotype segment information in the metadata
field.}

\item{minSNP, }{the crossover(s) will be filtered out if introduced by a segment
that has fewer than `minSNP` SNPs to support.}

\item{minlogllRatio, }{the crossover(s) will be filtered out if introduced by a
segment that has lower than `minlogllRatio` to its reversed state.}

\item{bpDist, }{the crossover(s) will be filtered out if introduced by a segment
that is shorter than `bpDist` basepairs.}

\item{maxRawCO, }{if a cell has more than `maxRawCO` number of raw crossovers
called across one chromosome, the cell is filtered out}
}
\value{
A `RangedSummarizedExperment` object that have different dims with input.
the colnames are the cell barcodes, rowRanges specify the location of SNPs that
contribute to crossovers.
}
\description{
This function filter out cells that have been called too many crossovers due
to diploid cell contamination or doublets. It also only keeps SNPs (rows) that
ever contribute to a crossover interval. This function should be run for 
individual chromosomes.
}
\details{
The `logllRatio` value is returned by `sscocaller` for each haplotype
segment formed by consecutive SNPs that are called to have a same state. It is
calculated by taking log of ratio (likelihood of SNPs with inferred states) 
and (likelihood of SNPs with reversed states)
}
\author{
Ruqian Lyu
}
