% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{findDupMarkers}
\alias{findDupMarkers}
\title{findDupMarkers}
\usage{
findDupMarkers(gt_matrix, threshold = 0.99, plot = TRUE)
}
\arguments{
\item{gt_matrix}{the corrected and filtered genotype matrix of markers by samples}

\item{threshold}{the frequency cut-off of number of matching genotypes out of
all geneotypes for determining whether the pair of samples
are duplicated, defaults to \code{0.9}}

\item{plot}{whether a frequency heatmap plot should be generated}
}
\description{
Find the duplicated markers by look at the number of matching genotypes across samples
}
