% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{findDupSamples}
\alias{findDupSamples}
\title{findDupSamples}
\usage{
findDupSamples(gt_matrix, threshold = 0.99, plot = TRUE)
}
\arguments{
\item{gt_matrix}{the corrected and filtered genotype matrix of markers by samples}

\item{threshold}{the frequency cut-off of number of matching genotypes out of
all geneotypes for determining whether the pair of samples
are duplicated, defaults to \code{0.99}}

\item{plot}{whether a frequency plot should be generated}
}
\description{
Find the duplicated samples by look at the number of matching genotypes
between all pairs of samples
}
