% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{getDistortedMarkers}
\alias{getDistortedMarkers}
\title{getDistortedMarkers}
\usage{
getDistortedMarkers(gt_matrix)
}
\arguments{
\item{gt_matrix}{the corrected and filtered genotype matrix of markers by samples}
}
\description{
Marker disortation detection using chisq-test
We expect the genotypes to appear with the frequenceis of 1:1 homo_alt:hets, we
use chisq-test for finding markers that have genotypes significantly diffferent from
the 1:1 ratio and report them
}
