% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{label_gt}
\alias{label_gt}
\title{`label_gt` for changing genotypes in alleles to labels}
\usage{
label_gt(s_gt, ref, alt, fail = "Fail")
}
\arguments{
\item{s_gt}{s_gt, a vector of genotypes for one sample across markers}

\item{ref}{ref, a vector of genotypes for reference strain across markers}

\item{alt}{alt, a vector of genotypes for alternative starin across markers}
}
\value{
a vector of labels \code{Homo_ref, Homo_alt, Het} indicating the progeny's
genotypes across markers
}
\description{
It turns a vector of Genotypes to a vector of Labels consist of
`Homo_ref`, `Homo_alt`, and `Het` given the genotypes for reference
and alternative strains.
}
\details{
This function takes the a sample's genotype across each SNP marker in alleles
and compare with genotypes of in-bred reference and alternative strains to. If the
sample's genotype for a particular SNP marker is the same with the reference
strain, it is labelled as Homo_ref \code{homogeneous reference} for a particular
SNP marker; if the sample's genotype is the same with the alternative strain
it is labelled as Homo_alt \code{homogeneous alternative} for a particular SNP
marker; if the sample's genotype is heterozygous then it is labelled as
Het \code{heterozygous} for this particular genotypes. If it does not fall in any
of the three cases, it is labelled as `missing`.

Note that the `Fail` genotype is still labelled as `Fail` after this function.
If there is a different label for failed genotype, provide the label using
the `fail` argument.
}
\author{
Ruqian Lyu
}
\keyword{internal}
