% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parents_geno}
\alias{parents_geno}
\title{Parents' genotype for F1 samples in `snp_geno`}
\format{
A data.frame:
\describe{
 \item{C57BL.6J}{genotype of reference mouse train across markers}
 \item{FVB.NJ..i.}{genotype of alternative mouse train across markers}
}
}
\usage{
parents_geno
}
\description{
Parents' genotype for F1 samples in `snp_geno`
}
\keyword{datasets}
