% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{permuteDist}
\alias{permuteDist}
\title{permuteDist}
\usage{
permuteDist(
  co_gr,
  B = 100,
  mapping_fun = "k",
  chr = NULL,
  group_prefix,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{co_gr}{GRanges the crossover counts for each marker interval across all
samples. Returned by \code{countCO}}

\item{B}{integer the number of sampling times}

\item{mapping_fun}{character default to "k" (kosambi mapping function). It 
can be one of the mapping functions: "k","h"}

\item{chr}{character the specific chromosome to test or NULL indicates all
chrs in the GRanges object will be summed together.}

\item{group_prefix}{a character vector of two}
}
\value{
the list of length B with numeric differences of two groups
}
\description{
Permutation test of two sample groups
}
\details{
It shuffles the group labels for the samples and calculate a difference
between two groups after shuffling.
}
\author{
Ruqian Lyu
}
