% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-count.R
\name{plotCount}
\alias{plotCount}
\alias{plotCount,RangedSummarizedExperiment,missing,missing-method}
\alias{plotCount,RangedSummarizedExperiment,character,missing-method}
\alias{plotCount,RangedSummarizedExperiment,character,logical-method}
\alias{plotCount,RangedSummarizedExperiment,missing,logical-method}
\alias{plotCount,GRanges,missing,logical-method}
\alias{plotCount,GRanges,missing,missing-method}
\alias{plotCount,GRanges,character,missing-method}
\alias{plotCount,GRanges,character,logical-method}
\title{plotCount}
\usage{
plotCount(co_count, group_by = "sampleGroup", by_chr = FALSE)

\S4method{plotCount}{RangedSummarizedExperiment,missing,missing}(co_count, group_by = "sampleGroup", by_chr = FALSE)

\S4method{plotCount}{RangedSummarizedExperiment,character,missing}(co_count, group_by = "sampleGroup", by_chr = FALSE)

\S4method{plotCount}{RangedSummarizedExperiment,character,logical}(co_count, group_by = "sampleGroup", by_chr = FALSE)

\S4method{plotCount}{RangedSummarizedExperiment,missing,logical}(co_count, group_by = "sampleGroup", by_chr = FALSE)

\S4method{plotCount}{GRanges,missing,logical}(co_count, group_by = "sampleGroup", by_chr = FALSE)

\S4method{plotCount}{GRanges,missing,missing}(co_count, group_by = "sampleGroup", by_chr = FALSE)

\S4method{plotCount}{GRanges,character,missing}(co_count, group_by = "sampleGroup", by_chr = FALSE)

\S4method{plotCount}{GRanges,character,logical}(co_count, group_by = "sampleGroup", by_chr = FALSE)
}
\arguments{
\item{co_count}{GRange or RangedSummarizedExperiment object, returned by \code{countCO}}

\item{group_by, }{the column name in `colData(co_count)` that specify the
grouping factor. Or the character vector contains the unique prefix of
sample names that are used for defining different sample groups. If missing
all samples are assumed to be from one group}

\item{by_chr, }{whether it should plot each chromosome separately}
}
\value{
ggplot object
}
\description{
Plot the number of COs per sample group or per chromosome
}
\examples{
demo_path <-paste0(system.file("extdata",package = "comapr"),"/")
s1_rse_state <- readHapState("s1",chroms=c("chr1"),
                             path=demo_path,barcodeFile=NULL,minSNP = 0,
                             minlogllRatio = 50,
                             bpDist = 100,maxRawCO=10,
                             minCellSNP = 0)
s1_count <- countCOs(s1_rse_state)
plotCount(s1_count)
}
