% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{plotGTFreq}
\alias{plotGTFreq}
\title{plotGTFreq}
\usage{
plotGTFreq(gt_matrix, interactive = FALSE)
}
\arguments{
\item{gt_matrix, }{the genotype matrix of marker by samples after correction by
funcion \code{correctGT}}

\item{plot, }{it determines whether a plot will be generated.}
}
\description{
Function to plot the genotypes for all samples faceted by genotype
}
\author{
Ruqian Lyu
}
