% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-genetic-dist.R
\name{plotWholeGenome}
\alias{plotWholeGenome}
\title{Plot cumulative genetic distances across the genome}
\usage{
plotWholeGenome(gr)
}
\arguments{
\item{gr, }{GRanges object with genetic distances calculated for marker
intervals}
}
\value{
A ggplot object
}
\description{
This function takes the calculated genetic distances for all
marker intervals across all chromosomes provided and plot the
cumulative genetic distances
}
\examples{
data(coCount)
dist_se <- calGeneticDist(coCount)
plotWholeGenome(SummarizedExperiment::rowRanges(dist_se))
}
