% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHapState.R
\name{readHapState}
\alias{readHapState}
\title{readHapState}
\usage{
readHapState(
  sampleName,
  chroms = c("chr1"),
  path,
  barcodeFile = NULL,
  minSNP = 30,
  minlogllRatio = 200,
  bpDist = 100,
  maxRawCO = 10
)
}
\arguments{
\item{sampleName, }{the name of the sample to parse which is used as prefix for
finding relevant files for the underlying sample}

\item{chroms, }{the character vectors of chromosomes to parse. Multiple chromosomes'
results will be concated together.}

\item{path, }{the path to the files, with name patterns *chrom_vi.mtx, *chrom_viSegInfo.txt,
end with slash}

\item{barcodeFile, }{if NULL, it is assumed to be in the same directory as the 
other files and with name sampleName_barcodes.txt}

\item{minSNP}{the crossover(s) will be filtered out if introduced by a segment
that has fewer than `minSNP` SNPs to support.}

\item{minlogllRatio}{the crossover(s) will be filtered out if introduced by a
segment that has lower than `minlogllRatio` to its reversed state.}

\item{bpDist}{the crossover(s) will be filtered out if introduced by a segment
that is shorter than `bpDist` basepairs.}

\item{maxRawCO}{if a cell has more than `maxRawCO` number of raw crossovers
called across one chromosome, the cell is filtered out}
}
\value{
a RangedSummarizedExperiment with rowRanges as SNP positions that contribute
to crossovers in any cells. colData contains cells annotation including barcodes and
sampleName.
}
\description{
A helper function that parses the viterbi state matrix (in .mtx format), 
barcode.txt and snpAnno.txt files for each individual.
}
\examples{
demo_path <-system.file("extdata",package = "comapr")
s1_rse_state <- readHapState(sampleName="s1",chroms=c("chr1"),
path=paste0(demo_path,"/"),
barcodeFile=NULL,minSNP = 0, minlogllRatio = 50,
bpDist = 100,maxRawCO=10)
s1_rse_state
}
\author{
Ruqian Lyu
}
