% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snp_geno}
\alias{snp_geno}
\title{Markers by genotype results for a group of samples}
\format{
A data frame with columns:
\describe{
 \item{C57BL.6J}{genotype of reference mouse train across markers}
 \item{FVB.NJ..i.}{genotype of alternative mouse train across markers}
 \item{POS}{SNP marker base-pair location }
 \item{CHR}{SNP marker chromosome location}
 \item{X100}{a mouse sample }
 \item{X101}{a mouse sample }
 \item{X102}{a mouse sample }
 \item{X103}{a mouse sample }
 \item{X104}{a mouse sample }
 \item{X105}{a mouse sample }
 \item{X106}{a mouse sample }
 \item{X107}{a mouse sample }
 \item{X108}{a mouse sample }
 \item{X109}{a mouse sample }
 \item{X110}{a mouse sample }
 \item{X111}{a mouse sample }
 \item{X112}{a mouse sample }
 \item{X113}{a mouse sample }
 \item{X92}{a mouse sample }
 \item{X93}{a mouse sample }
 \item{X94}{a mouse sample }
 \item{X95}{a mouse sample }
 \item{X96}{a mouse sample }
 \item{X97}{a mouse sample }
 \item{X98}{a mouse sample }
 \item{X99}{a mouse sample }
 \item{rsID}{the SNP ID}
}
}
\source{
Statistics Canada. Table 001-0008 - Production and farm value of
 maple products, annual. \url{http://www5.statcan.gc.ca/cansim/}
}
\usage{
data(snp_geno)
}
\description{
Markers by genotype results for a group of samples
}
\keyword{datasets}
