% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snp_geno_gr}
\alias{snp_geno_gr}
\title{Markers by genotype results for a group of samples}
\format{
A GRanges object:
\describe{
 \item{X100}{a mouse sample }
 \item{X101}{a mouse sample }
 \item{X102}{a mouse sample }
 \item{X103}{a mouse sample }
 \item{X104}{a mouse sample }
 \item{X105}{a mouse sample }
 \item{X106}{a mouse sample }
 \item{X107}{a mouse sample }
 \item{X108}{a mouse sample }
 \item{X109}{a mouse sample }
 \item{X110}{a mouse sample }
 \item{X111}{a mouse sample }
 \item{X112}{a mouse sample }
 \item{X113}{a mouse sample }
 \item{X92}{a mouse sample }
 \item{X93}{a mouse sample }
 \item{X94}{a mouse sample }
 \item{X95}{a mouse sample }
 \item{X96}{a mouse sample }
 \item{X97}{a mouse sample }
 \item{X98}{a mouse sample }
 \item{X99}{a mouse sample }
 \item{rsID}{the SNP ID}
}
}
\source{
TBD
}
\usage{
snp_geno_gr
}
\description{
Markers by genotype results for a group of samples
}
\examples{
 snp_geno_gr

}
\keyword{datasets}
