% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twoSamples}
\alias{twoSamples}
\title{RangedSummarizedExperiment object containing the Viterbi states SNP markers
for samples from two groups. `colData(twoSamples)` contains the sample group
factor.}
\format{
An object of class \code{RangedSummarizedExperiment} with 6 rows and 10 columns.
}
\usage{
twoSamples
}
\description{
RangedSummarizedExperiment object containing the Viterbi states SNP markers
for samples from two groups. `colData(twoSamples)` contains the sample group
factor.
}
\keyword{datasets}
