% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{correctGT}
\alias{correctGT}
\title{correctGT}
\usage{
correctGT(gt_matrix, ref, alt)
}
\arguments{
\item{gt_matrix}{the input from excel with rownames as marker IDs and column names as sample IDs}

\item{ref}{is the reference genotype}

\item{alt}{is the alternative genotype.}
}
\description{
map GT to genotype labels, change Homo_ref to Hets, infer missing data,
and change "missing" to NA for counting crossover later
}
\details{
This function changes genotype code to labels by calling \code{lable_gt}, corrects
the Home reference geneotype calls to Hets by calling \code{correct_ref}, infer the
failed genotype calls by calling \code{fill_fail} and change missing values to NA by
calling \code{change_missing}.
}
\author{
Ruqian Lyu
}
