% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count-co.R
\name{countCOs}
\alias{countCOs}
\title{countCOs}
\usage{
countCOs(geno, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{geno}{GRanges object with SNP positions and genotypes codes}

\item{BPPARAM}{the registered backend for parallel computing
across samples}
}
\value{
data.frame with markers as rows and samples in columns, values as the
number of COs estimated for each marker interval
}
\description{
Count number of COs within each marker interval
COs identified in the interval overlapping missing markers are distributed
according to marker interval base-pair sizes
}
\author{
Ruqian Lyu
}
