% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{fill_fail}
\alias{fill_fail}
\title{Function fill in FAIL
If we have a \code{FAIL} in the genotype data and the \code{FAIL} in a block of either Home_alt,
or Het, we fill in the \code{FAIL} using values of the ones adjacent to it,
otherwise we fill in "NA" as missing value.}
\usage{
fill_fail(s_gt, fail = "Fail")
}
\arguments{
\item{s_gt, }{a column of labelled genotypes}

\item{fail, }{the string that is used for encoding failed genotype results, default to \code{Fail}}
}
\description{
Function fill in FAIL
If we have a \code{FAIL} in the genotype data and the \code{FAIL} in a block of either Home_alt,
or Het, we fill in the \code{FAIL} using values of the ones adjacent to it,
otherwise we fill in "NA" as missing value.
}
\keyword{internal}
