% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{findDupMarkers}
\alias{findDupMarkers}
\title{findDupMarkers}
\usage{
findDupMarkers(geno, threshold = 0.99, plot = TRUE, in_text = TRUE, ...)
}
\arguments{
\item{geno}{the genotype data.frame of markers by samples from output of
function \code{correctGT}}

\item{threshold}{the frequency cut-off for determining whether the pair of markers
are duplicated, defaults to \code{0.9}}

\item{plot}{whether a frequency heatmap plot should be generated}
}
\value{
The paris of duplicated markers.
}
\description{
Find the duplicated markers by look at the number of matching genotypes across samples
}
