% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-af-co.R
\name{getCellCORange}
\alias{getCellCORange}
\title{getCellCORange}
\usage{
getCellCORange(co_count, cellBarcode)
}
\arguments{
\item{co_count, }{`GRange` or `RangedSummarizedExperiment` object,}

\item{cellBarcode, }{the selected cell's barcode}
}
\value{
GRange object containing the crossover intervals for the selected
cell
}
\description{
It finds the crossover intervals for a selected cell
}
\author{
Ruqian Lyu
}
