% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-distorted-markers.R
\name{getDistortedMarkers}
\alias{getDistortedMarkers}
\title{getDistortedMarkers}
\usage{
getDistortedMarkers(geno, p = c(0.5, 0.5))
}
\arguments{
\item{geno}{the genotype data.frame of markers by samples from output of
function \code{correctGT}}

\item{p}{the expected geneotype ratio in a numeric vector, defauls to c(0.5,0.5)}
}
\description{
Marker disortation detection using chisq-test
}
\details{
We expect the genotypes to appear with the frequenceis of 1:1 homo_alt:hets.
We use chisq-test for finding markers that have genotypes among samples that
are significantly diffferent from the 1:1 ratio and report them
}
\author{
Ruqian Lyu
}
