% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{label_gt}
\alias{label_gt}
\title{Function label_gt for changing genotypes to labels}
\usage{
label_gt(s_gt, ref, alt, fail = "Fail")
}
\arguments{
\item{s_gt}{s_gt, a vector of genotypes for one sample across markers}

\item{ref}{ref, a vector of genotypes for reference across markers}

\item{alt}{alt, a vector of genotypes for alternative across markers}
}
\value{
a vector of labels indicating the progeny's genotypes across markers
: Homo_ref, Homo_alt, Het
}
\description{
It turns a vector of Genotypes to Labels: Homo_ref, Homo_alt, Het
given the known reference and alternative genotypes
}
\details{
This function takes the a sample's genotype vector and checks for each SNP marker
to see whether it matches with homogeneous reference, or homogeneous alternative
or heterozygous genotypes.
}
\author{
Ruqian Lyu
}
\keyword{internal}
