% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rb.R
\name{plotMissingGT}
\alias{plotMissingGT}
\title{Plot markers with missing genotypes}
\usage{
plotMissingGT(gt_matrix, missing = "Fail", plot_wg = TRUE)
}
\arguments{
\item{gt_matrix}{the genotype matrix of markers by samples}

\item{missing}{the label in the matrix that is used for encoding the missing or failed data}

\item{plot_wg}{whether to plot all markers across whole genome or just the missing markers}
}
\value{
a plot for markers with missing genotype across samples
}
\description{
Plot markers with missing genotypes
}
\author{
Ruqian Lyu
}
